KIX2018-6473: consider relevant organisation id

See merge request Softwareentwicklung/KIXng/backend/app!1119

+++ Kernel/System/Contact.pm
@@ -880,8 +880,8 @@ sub ContactUpdate {
     # update organisation IDs
     if (@DeleteOrgIDs) {
         return if !$Kernel::OM->Get('DB')->Do(
-            SQL  => 'DELETE FROM contact_organisation WHERE contact_id = ? AND org_id IN (?)',
-            Bind => [ \$Param{ID}, \${\(join ', ', @DeleteOrgIDs)} ],
+            SQL  => "DELETE FROM contact_organisation WHERE contact_id = ? AND org_id IN ( ${\(join ', ', @DeleteOrgIDs)} )",
+            Bind => [ \$Param{ID} ],
         );
     }
